+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+- Enable option to Manage Changelog
+- Enable setting Labels on PRs created
+
 ## 2.0.0 - *2021-08-11*
 
 - Breaking Change: Defaults to looking for main branches, introduces new cli arg for setting default branch name